1. 編輯 app/User.php，加入三種帳號角色名稱的常數字串，並將 role 這個欄位加入 fillable 中。 2. 編輯 database/migrations/2014_10_12_000000_create_users_table.php 這一個建立使用者資料表的 migration 設定檔，在其中加入一欄用於儲存帳號角色的 role 欄位。 3. 編輯 app/Http/Controllers/Auth/RegisterController.php，設定使用者註冊時的角色都是一般使用者。 4. 建立 Gate 權限規則。 5.自動生成三個 users: admin, manager, and user (方便我們做測試之用) 6.更新兩個views： records.index ,keys.index,usernames.index讓他們針對admin 與 manager 呈現受控管的超連結(新增、修改、刪除) 7.讓每一個路由都必須經過 auth 的 middleware (即必須登入帳號才可以使用)。 8.增加 can 的middleware，讓新增表單、儲存資料、刪除資料路由只能讓admin使用 9.不允許編輯表單給 user 權限使用，但允許 admin或manager使用

+++ routes/web.php
@@ -4,6 +4,7 @@ use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\RecordsController;
 use App\Http\Controllers\KeysController;
 use App\Http\Controllers\UsernamesController;
+use Illuminate\Support\Facades\Auth;
 /*
 |--------------------------------------------------------------------------
 | Web Routes
@@ -14,6 +15,7 @@ use App\Http\Controllers\UsernamesController;
 | contains the "web" middleware group. Now create something great!
 |
 */
+Route::middleware(['auth'])->group(function () {
 Route::get('/',function(){
     return redirect('records');
 });
@@ -24,13 +26,13 @@ Route::get('records/{id}', [RecordsController::class,'show'])->where('id', '[0-9
 // 修改單一鑰匙借用紀錄表單
 Route::get('records/{id}/edit', [RecordsController::class,'edit'])->where('id', '[0-9]+')->name('records.edit');
 // 刪除單一鑰匙借用紀錄資料
-Route::delete('records/delete/{id}',[RecordsController::class,'destroy'])->where('id', '[0-9]+')->name('records.destroy');
+Route::delete('records/delete/{id}',[RecordsController::class,'destroy'])->where('id', '[0-9]+')->name('records.destroy')->middleware('can:admin');
 // 新增鑰匙借用紀錄表單
-Route::get('records/create', [RecordsController::class, 'create'])->name('records.create');
+Route::get('records/create', [RecordsController::class, 'create'])->name('records.create')->middleware('can:admin');
 // 修改鑰匙借用紀錄資料
 Route::patch('records/update/{id}', [RecordsController::class, 'update'])->where('id', '[0-9]+')->name('records.update');
 // 儲存鑰匙借用紀錄資料
-Route::post('records/store', [RecordsController::class, 'store'])->where('id', '[0-9]+')->name('records.store');
+Route::post('records/store', [RecordsController::class, 'store'])->where('id', '[0-9]+')->name('records.store')->middleware('can:admin');
 //2023年借用查詢
 Route::get('records/borrow',[RecordsController::class,'borrow'])->name('records.borrow');
 //選定鑰匙查詢借用者
@@ -43,13 +45,13 @@ Route::get('keys/{id}', [KeysController::class,'show'])->where('id', '[0-9]+')->
 // 修改單一鑰匙編號和地點表單
 Route::get('keys/{id}/edit', [KeysController::class,'edit'])->where('id', '[0-9]+')->name('keys.edit');
 // 刪除單一鑰匙編號和地點資料
-Route::delete('keys/delete/{id}',[KeysController::class,'destroy'])->where('id', '[0-9]+')->name('keys.destroy');
+Route::delete('keys/delete/{id}',[KeysController::class,'destroy'])->where('id', '[0-9]+')->name('keys.destroy')->middleware('can:admin');
 // 新增鑰匙編號和地點表單
-Route::get('keys/create', [KeysController::class, 'create'])->name('keys.create');
+Route::get('keys/create', [KeysController::class, 'create'])->name('keys.create')->middleware('can:admin');
 // 修改鑰匙編號和地點資料
 Route::patch('keys/update/{id}', [KeysController::class, 'update'])->where('id', '[0-9]+')->name('keys.update');
 // 儲存鑰匙編號和地點資料
-Route::post('keys/store', [KeysController::class, 'store'])->where('id', '[0-9]+')->name('keys.store');
+Route::post('keys/store', [KeysController::class, 'store'])->where('id', '[0-9]+')->name('keys.store')->middleware('can:admin');
 
 //顯示所有教職員生
 Route::get('usernames',[UsernamesController::class,'index'])->name('usernames.index');
@@ -58,10 +60,14 @@ Route::get('usernames/{id}', [UsernamesController::class,'show'])->where('id', '
 // 修改單一教職員生表單
 Route::get('usernames/{id}/edit', [UsernamesController::class,'edit'])->where('id', '[0-9]+')->name('usernames.edit');
 // 刪除單一教職員生資料
-Route::delete('usernames/delete/{id}',[UsernamesController::class,'destroy'])->where('id', '[0-9]+')->name('usernames.destroy');
+Route::delete('usernames/delete/{id}',[UsernamesController::class,'destroy'])->where('id', '[0-9]+')->name('usernames.destroy')->middleware('can:admin');
 // 新增教職員生表單
-Route::get('usernames/create', [UsernamesController::class, 'create'])->name('usernames.create');
+Route::get('usernames/create', [UsernamesController::class, 'create'])->name('usernames.create')->middleware('can:admin');
 // 修改教職員生資料
 Route::patch('usernames/update/{id}', [UsernamesController::class, 'update'])->where('id', '[0-9]+')->name('usernames.update');
 // 儲存教職員生資料
-Route::post('usernames/store', [UsernamesController::class, 'store'])->where('id', '[0-9]+')->name('usernames.store');
+Route::post('usernames/store', [UsernamesController::class, 'store'])->where('id', '[0-9]+')->name('usernames.store')->middleware('can:admin');
+});
+Auth::routes();
+
+Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
